Implementacao da remocao de aluno em modo offline - soft delete

+++ app/src/main/java/br/com/alura/agenda/dao/AlunoDAO.java
@@ -20,7 +20,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
 
 
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -33,7 +33,8 @@ public class AlunoDAO extends SQLiteOpenHelper {
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT, " +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
 
         db.execSQL(sql);
     }
@@ -88,6 +89,9 @@ public class AlunoDAO extends SQLiteOpenHelper {
             case 4:
                 String adicionaCampoSincronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSincronizado);
+            case 5:
+                String adicionaCampoDesativado = "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0";
+                db.execSQL(adicionaCampoDesativado);
 
 
         }
@@ -115,7 +119,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
@@ -137,6 +141,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
 
             alunos.add(aluno);
         }
@@ -155,13 +160,20 @@ public class AlunoDAO extends SQLiteOpenHelper {
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
+
         return dados;
     }
 
     public void deleta(Aluno aluno) {
         SQLiteDatabase db = getWritableDatabase();
         String [] params = {aluno.getId().toString()};
-        db.delete("Alunos", "id = ?", params);
+        if(aluno.estaDesativado()) {
+            db.delete("Alunos", "id = ?", params);
+        }else{
+            aluno.desativar();
+            edita(aluno);
+        }
     }
 
     public void edita(Aluno aluno) {
